Add all data in Calculator.js
